Refact: altera tipo do id do usuário para Long

+++ src/main/java/br/com/transportae/usuario/IUsuarioRepository.java
@@ -1,15 +1,14 @@
 package br.com.transportae.usuario;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 
-public interface IUsuarioRepository extends JpaRepository<UsuarioModel, BigInteger> {
+public interface IUsuarioRepository extends JpaRepository<UsuarioModel, Long> {
     
-    Optional<UsuarioModel> findById(BigInteger id);
+    Optional<UsuarioModel> findById(Long id);
     Optional<UsuarioModel> findByCpf(String cpf);
     Optional<UsuarioModel> findByEmail(String email);
